Proof of concept: Add support to general system policy to authorize by an access token
Still a WIP - A token can be issued for a particular actor to perform a specified set of operations on a specified resource.  If the token is included in the helper policy call, then the TokenOperation will pick it up, validate the token, and permit access.  Alterately, a persisted token could be queried for the specified actor (code not currently in place), which would allow a person to grant specific access to a user via a token versus via an entitlement.