Add join clause
This is still rudimentary, allowing for models to be listed while joined to another table, such as participation, in order to list out the participations (containers) or participants (members) with optional effect.  In previous versions and the initial A7 prototype, the participations were queried first, and then the records reconstituted from those results.