Add support to arbitrarily encrypt fields
A field designated as being encrypted will be enciphered with a rotating encrypted vault key associated with the organization's vaultuser.  Currently this will work for byte arrays and string values.  Note: field dimensions need to auto-scale to support an encrypted value which would likely exceed the maximum length limit of an unencrypted value