Add support to stream-in-place
This effectively replaces the previous version 'pointer' support.  This allows a data object to refer to a file, but only when the object was loaded/updated via the console, and only when the web service parameter enabled the feature.  Obviously there were a lot of security concerns with that setup!  In the new version, supported path roots must be pre-configured and which affect both update and read.  Also, the stream object is meant to represent multiple sources beyond file-based, so as needed other data sources may easily be added to the stream reader and writer, with all of the other authorization policies still applying.