Add optional container-first policy check to bulk updates
When invoking update, or at the class level, bulk update operations will first check for a container (group or parent id) and perform a coarse policy check to whether the user can update it.  That result will be used for every subsequent object with that container id, instead of performing the object level policy check, which more aggressively validates to the field and foreign reference level.  This option could allow foreign references to incorrectly be updated  for bulk updates - the trade off is the authorization check is a lot faster and useful when loading large sets of data.