Add utilities to cleanup orphaned foreign reference
At the moment, and for every prior version, foreign constraints were not used because there were never uniform rules (except perhaps with groups and users) as to how to deal with the broken foreign keys, so a lazy-cascade was used to clean it up at a later time.  This implementation is a more streamlined implementation of that lazy cascade, though I ought to add a foreign key constraint option at some point.