Raw notes on condensed action format
The action format combines select verbs (via wordnet) with personality, statistic, and state annotation, coupled with an opposing verb, and/or follow-on verb, such that a basic situation may be expressed in verbs with related statistics, relative to any parent state, and the actionResponse used to keep track of outcomes.

One item I've been meaning to add is a condensed JSON format, where in a model can be minified to one or two character properties in order to more easily transport and edit many similar records at once.