Add option to follow single foreign references from the top level
The initial behavior was to only include a model name and the identifier, and then require an additional call to populate that field.  However, with the embedded syntax used for the lists, this is swapped out to allow the model to be populated with common fields are, when the limiter is disabled, all fields (except blobs).