Incremental update to advancing increments
The general idea is that the rules can define what an 'increment' is.  The default is year/month/day/hour, but if the preference is to keep a running clock or by year, then a custom rule can be added, such that subsequent rules only pick up on a) the passage of time, b) the prior state, c) any schedules that should have been followed, and d) an evaluation of what the current state should be.