Fix issue deserializing complex lists with condensed model declarations
Condensing the model name allows for the model name to be skipped after the first item in a list, and skipped for foreign child references.  However, the previous model name was not being correctly  used when a child array overwrote the marker. Now the model marker is sensitive to the depth of deserialization.