Load item templates into the world instead of the universe
Also, mark items cloned off animals templates as not being templates since these are meant to be discrete entries.

The whole item assignment needs to be redone since it's very verbose and should really just be an inventory of counts and qualities with pointers back to the template, so instead of cloning the item, define a new inventory item that can optionally override the quality but for most things can be a simple pointer with age and consumption tracking.