Add inventory dimension to represent item instances
While items may be directly associated with another object for purposes of reflecting degredation and modification, other items that remain largely unchanged such as consumables may be referenced as an inventory item vs direct item assignment.  Yes, this could be handled at the item level as well .... but, I didn't want to.