Add initial rules and calculations for personality based contests
For example, when identifying a leader in a sub group of people with leadership tendencies, evaluate against maturity, psychopathy, narcissism, and machiavellianism - note that the counter statistics are a bit fudged at the moment, usually averaging counter personality traits.  Also, the leadership contest recurses less the assumed leader if that leader loses a personality contest.