Add option to decompress/decrypt databytestore field on serialization
This is specific to the databytestore field vs encrypted fields.  This particular field is a holdover from the original design as it combines compression, vaulting, and enciphering together, versus the newer design of adding vaulted enciphering at the field level.  Field level compression isn't yet defined, so when it is then this can be backed out.  Until then, this setting allows the data to be serialized for consumption (eg: to an API consumer) vs serialized for persistence where the encryption and compression should apply.