Choose interactionDescription over scene
The hard coded prompt to create the interaction description produces a more robust description, while the scene description was intentionally meant to be a brief statement.  Currently experimenting to see which result is more favorable - note that all generated descriptions are preserved on the chatConfig and user and system narrative objects for each, so that each config has a stateful copy of the generated scene and setting.