Include vault fields in vector search
The abstract VaultExt model uses the EncryptField provider to encrypt the plain text content.  This means that a) the query must include the vaultid, vaulted bit, and keyid  fields, and b) the MemoryReader must be used on the reconstituted model in order to decrypt the content field.  This is a special case in that the vector query is still experimental and doesn't use the common io.StatementUti class to performa model operations against the database.