Attach the final summary as a vector store to the reference model
The summary sets are stored and vectorized in an adjacent object at the moment (which I don't like, but I do like having all of the content available).  By adding the summary of summaries to the reference vector store, then it's available when citing only the reference and not having to specifically cite the reference summary.